requirements.txt
amqp==5.2.0
APScheduler==3.10.4
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.3
django-allauth==0.61.1
django-apscheduler==0.6.2
django-filter==24.2
django-modeltranslation==0.19.3
djangorestframework==3.15.2
dnspython==2.6.1
eventlet==0.36.1
flake8==7.0.0
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
kombu==5.3.7
Markdown==3.6
mccabe==0.7.0
oauthlib==3.2.2
packaging==24.0
pluggy==1.5.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.8.0
pytest==8.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
redis==5.0.4
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
